[Refactor] Cart 리팩토링 및 버그 수정
