+++ src/components/Header/Cart/Cart.js
@@ -3,131 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import './Cart.scss';
 import Modal from 'react-modal';
 import axios from 'axios';
-
-function ItemBox({
-  setItemData,
-  cartId,
-  name,
-  img,
-  price,
-  num,
-  productId,
-  setIsCartClicked,
-}) {
-  const navigate = useNavigate();
-
-  const scrollToTop = () => {
-    window.scrollTo({
-      top: 0,
-      behavior: 'auto',
-    });
-  };
-
-  function plus() {
-    axios({
-      method: 'put',
-      url: 'http://localhost:8000/cart/',
-      data: {
-        token: localStorage.getItem('token'),
-        cart_id: cartId,
-        num: num + 1,
-      },
-    }).then(res => {
-      setItemData(res.data[0]);
-    });
-  }
-
-  function minus() {
-    axios({
-      method: 'put',
-      url: 'http://localhost:8000/cart/',
-      data: {
-        token: localStorage.getItem('token'),
-        cart_id: cartId,
-        num: num - 1,
-      },
-    }).then(res => {
-      if (res.data.length === 0)
-        setItemData({
-          cart_id: [],
-          product_id: [],
-          product_name: [],
-          product_price: [],
-          num: [],
-          product_photos: [],
-        });
-      else setItemData(res.data[0]);
-    });
-  }
-
-  function remove() {
-    axios({
-      method: 'delete',
-      url: 'http://localhost:8000/cart',
-      data: {
-        token: localStorage.getItem('token'),
-        cart_id: cartId,
-      },
-    }).then(res => {
-      if (res.data.length === 0)
-        setItemData({
-          cart_id: [],
-          product_id: [],
-          product_name: [],
-          product_price: [],
-          num: [],
-          product_photos: [],
-        });
-      else setItemData(res.data[0]);
-    });
-  }
-
-  return (
-    <div className="item_box">
-      <img src={img} alt="item" className="pic" />
-      <div className="text_box">
-        <div
-          className="title"
-          onClick={() => {
-            navigate(`/product/detail/${productId}`);
-            scrollToTop();
-            setIsCartClicked(false);
-          }}
-        >
-          {name}
-        </div>
-        <div className="price">$ {price.toLocaleString()}</div>
-        <div className="wrapper">
-          <img
-            src="/image/add.png"
-            alt="add"
-            className="add"
-            onClick={() => {
-              plus();
-            }}
-          />
-          <p className="quantity">{num}</p>
-          <img
-            src="/image/minus.png"
-            alt="minus"
-            className="minus"
-            onClick={() => {
-              minus();
-            }}
-          />
-          <div
-            className="remove"
-            onClick={() => {
-              remove();
-            }}
-          >
-            Remove
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
+import CartItemBox from '../../../pages/CartDetail/CartItemBox/CartItemBox';
 
 function Cart({ setIsCartClicked, setScrollPosition }) {
   const navigate = useNavigate();
@@ -182,7 +58,7 @@ function Cart({ setIsCartClicked, setScrollPosition }) {
           className="buy_btn flex_center"
           onClick={() => {
             setIsCartClicked(false);
-            moveAndScrollToTop('/shop');
+            moveAndScrollToTop('/shop/1');
           }}
         >
           SHOP
@@ -216,7 +92,7 @@ function Cart({ setIsCartClicked, setScrollPosition }) {
         itemData.num.length !== 0 ? (
           itemData.product_name.map((el, i) => {
             return (
-              <ItemBox
+              <CartItemBox
                 cartId={itemData.cart_id[i]}
                 productId={itemData.product_id[i]}
                 key={itemData.product_id[i]}
@@ -252,4 +128,4 @@ function Cart({ setIsCartClicked, setScrollPosition }) {
   );
 }
 
-export { Cart, ItemBox };
+export default Cart;
