+++ src/components/Header/ExpandHeader/ExpandHeader.js
@@ -5,15 +5,15 @@ import ItemContainer from '../../../pages/Home/ItemContainer/ItemContainer';
 import { useEffect, useRef, useState } from 'react';
 
 const trend = [
-  { url: '/shop?category=gift', name: 'Gift Set' },
-  { url: '/shop?category=limited', name: 'Limited Edition' },
-  { url: '/shop?category=bars', name: 'Bars' },
+  { url: '/shop/1?category=gift', name: 'Gift Set' },
+  { url: '/shop/1?category=limited', name: 'Limited Edition' },
+  { url: '/shop/1?category=bars', name: 'Bars' },
 ];
 
 const popular = [
-  { url: '/shop?category=pure', name: 'Pure Chocolate' },
-  { url: '/shop?category=sweet', name: 'Sweet & Salty' },
-  { url: '/shop?category=wafers', name: 'Wafers' },
+  { url: '/shop/1?category=pure', name: 'Pure Chocolate' },
+  { url: '/shop/1?category=sweet', name: 'Sweet & Salty' },
+  { url: '/shop/1?category=wafers', name: 'Wafers' },
 ];
 
 export default function ExpandHeader({ isShopClicked, setIsShopClicked }) {
